Map ids are ints in 2.4

+++ core/src/main/scala/org/apache/spark/shuffle/ShuffleBlockInfo.scala
@@ -25,4 +25,4 @@ import org.apache.spark.annotation.Experimental
  * type safe way.
  */
 @Experimental
-case class ShuffleBlockInfo(shuffleId: Int, mapId: Long)
+case class ShuffleBlockInfo(shuffleId: Int, mapId: Int)
